feat: support ts import `type` modifiers on import names
This feature is introduced in TypeScript 4.5:
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-5.html#type-modifiers-on-import-names

And recommended in `@vue/tsconfig`:
https://github.com/vuejs/tsconfig/blob/26fc528db347dbbbbaa2ea15d3c2d95c197887b7/tsconfig.json#L19-L23

Sucrase does not support it yet:
https://github.com/alangpierce/sucrase/issues/668

So I replaced it with `@babel/standalone`.
Babel supported the new syntax since v7.16.
So I also updated the project dependencies to ensure that
`vue/compiler-sfc` can also parse this syntax.

The bundle is a lot bigger after this change:
```
dist/vue-repl.js   5040.05 KiB / gzip: 840.42 KiB
```

But it is still much smaller than the size of a full TypeScript compiler
or `esbuild-wasm`, which are 9MiB+ in size.