fix(codemirror): fix codemirror editor showing nothing on start on small layouts when starting in ouput mode
CodeMirror Editor tries to auto-size when it starts.
If the Repl start in Output mode on small layouts CodeMirror will not be able to autosize because it has `display: none`.
From the POV of user the only solution is to change editor tabs to see the code (if it has more tabs).